Updater: Set SYSTEM_UPDATE_SETTINGS priority higher than GMS
For whatever reason, when both SYSTEM_UPDATE_SETTINGS components have
the same priority and the first one is disabled (in this case, the GMS
one), it doesn't seem to fallback to the next enabled one and instead
shows a "No apps can perform this action" message. To avoid that, let's
just set higher priority.

9a6853 com.google.android.gms/.update.SystemUpdateActivity filter b9e590
  Action: "android.settings.SYSTEM_UPDATE_SETTINGS"
  Category: "android.intent.category.DEFAULT"
  mPriority=1, mOrder=0, mHasStaticPartialTypes=false, mHasDynamicPartialTypes=false

Change-Id: I886a714304fe6d1906af7e1702ecaf2238bd8cb0
Signed-off-by: Aditya Eka Pranata <aepranata@gmail.com>